Update footer and separate about and contact sections
